requirements.txt
azure-ai-documentintelligence==1.0.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.20.0
azure-search-documents==11.4.0
azure-storage-blob==12.24.1
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b8
fastapi==0.115.8
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
numpy==1.26.4
openai==1.61.0
urllib3==2.3.0
uvicorn==0.34.0
